outgoing SessionMessage management
add SessionMessageSender class to wrap a queue of SessionMessages to be
sent to a given identifier.

When SessionManager sees an identifier, a new SessionMessageSender is
created for it and initiated with an outgoing IdentityMessage. As
clients request SessionMessages be sent to Peers, the SessionManager
adds those SessionMessages to the appropriate SessionMessageSender.